Split the medication frequency handler into 2 parts to create a basereminder
